# API/Infernum-API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Infernum API",
        "description": "Documentación de la API Infernum",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:1606/api",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/v1/profile/show": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get authenticated user profile (TOKEN REQUIRED)",
                "description": "**NO TOKEN = 401!**\\n\\n1. **POST /v1/login** → copy token\\n2. **Authorize** → paste `Bearer {token}`. **Execute this route** ✅",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "User OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfileResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "TOKEN required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error: Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error: Failure"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update authenticated user profile (TOKEN REQUIRED)",
                "description": "**NO TOKEN = 401!**\\n\\n1. **POST /v1/login** → copy token\\n2. **Authorize** → paste `Bearer {token}`. **Execute this route** ✅",
                "operationId": "updateProfile",
                "requestBody": {
                    "description": "Data of the profile updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "TOKEN required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error: Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error: Failure"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/cart/show": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get shopping cart",
                "operationId": "showShoppingCart",
                "responses": {
                    "200": {
                        "description": "Shopping cart successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Succesfull"
                                        },
                                        "cart": {
                                            "$ref": "#/components/schemas/ShoppingCartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not have a shopping cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failure: The user does not have a shopping cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/cart/create": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create shopping cart or add/update item",
                "operationId": "createCartItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShoppingCartRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item created or quantity successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Succesfull"
                                        },
                                        "message": {
                                            "type": "string",
                                            "enum": [
                                                "quantity updated",
                                                "Cart and item created",
                                                "item added"
                                            ]
                                        },
                                        "cart_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "game_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You need to send the game_id"
                                                    }
                                                },
                                                "quantity": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You need to introduce a quantity"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateShoppingCartRequest": {
                "title": "CreateShoppingCartRequest",
                "description": "Request to create or add item to shopping cart",
                "required": [
                    "game_id"
                ],
                "properties": {
                    "game_id": {
                        "description": "Game ID (must exist in games table)",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ProfileRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "display_name": {
                        "type": "string",
                        "example": "manolo"
                    },
                    "bio": {
                        "type": "text",
                        "example": "I am just a player"
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": "https://image.webp"
                    }
                },
                "type": "object"
            },
...
            "ProfileResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "display_name": {
                        "type": "string",
                        "example": "example_user"
                    },
                    "bio": {
                        "type": "string",
                        "example": "I am a normal user"
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": "https://png.webp"
                    }
                },
                "type": "object"
            },
...
            "ShoppingCartResource": {
                "title": "ShoppingCartResource",
                "description": "Shopping cart resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "$ref": "#components/schemas/CartItemsResource"
                    }
                },
                "type": "object"
            },
            "UserProfileResource": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/ProfileResource"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "john_doe"
                    },
                    "role": {
                        "type": "string",
                        "example": "client"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Ingresa tu token de Sanctum",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}